# Postman/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "GestaoColaboradores API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Colaboradores/{id}": {
      "get": {
        "tags": [
          "Colaboradores"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Colaborador"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Colaborador"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Colaborador"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Colaboradores"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Colaborador"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Colaborador"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Colaborador"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Colaboradores"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Usuarios": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Usuario"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Usuario"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Usuario"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Usuarios"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Usuario"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Usuario"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Usuario"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          }
        }
      }
    },
    "/api/Usuarios/{id}": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Usuarios"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Usuario"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Usuario"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Usuario"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Usuarios"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Colaborador": {
        "type": "object",
        "properties": {
          "idColaborador": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "nome": {
            "type": "string",
            "nullable": true
          },
          "codigoUnidade": {
            "type": "integer",
            "format": "int32"
          },
          "idUsuario": {
            "type": "integer",
            "format": "int32"
          },
          "unidade": {
            "$ref": "#/components/schemas/Unidade"
          }
        },
        "additionalProperties": false
      },
      "Unidade": {
        "type": "object",
        "properties": {
          "idUnidade": {
            "type": "integer",
            "format": "int32"
          },
          "codigoUnidade": {
            "type": "integer",
            "format": "int32"
          },
          "nomeUnidade": {
            "type": "string",
            "nullable": true
          },
          "unidadeAtiva": {
            "type": "boolean"
          },
          "colaboradores": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Colaborador"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Usuario": {
        "type": "object",
        "properties": {
          "idUsuario": {
            "type": "integer",
            "format": "int32"
          },
          "nomeLogin": {
            "type": "string",
            "nullable": true
          },
          "senha": {
            "type": "string",
            "nullable": true
          },
          "usuarioAtivo": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}